tasks/requirements.txt
cookiecutter
future
git-aggregator
invoke
pyyaml
requests
ruamel.yaml
gnupg
psycopg2

# marabunta new versioning
marabunta >= 0.9.0

